Add python file with 3 functions
The python directory is intended to contain the implementation of the algorithms
written in Python.

karatsuba.py implements the karatsuba algorithm, which is used to multiply
two integers.

merge_sorte.py is used to sorte lists using the merge and sort algorithm.

rec_int_mult.py calculates the multiplication between two integers. It uses a
recursive algorithm for multiplying both numbers.